FISH-7503: split mojo class and increment test coverage
