+++ README.md
@@ -24,13 +24,13 @@ mvn clean install
 Execute Command under a project that you want to advise for jakarta 10
 
 ```
-mvn fish.payara.advisor:advisor-maven-plugin:1.0-SNAPSHOT:advice
+mvn fish.payara.advisor:advisor-maven-plugin:1.0-SNAPSHOT:advise
 ```
 
 if a pattern matchs you will see something like the following:
 
 ```
-[INFO] Showing Advices
+[INFO] Showing advises
 [INFO] ********
  Jakarta Authorization 2.1
  Issue # 105
