+++ src/main/resources/config/jakarta10/advisorFix/jakarta-restful-ws-fix-messages.properties
@@ -41,12 +41,12 @@ jakarta-restful-ws-method-change-must-favor-jsonb=This issue won't affect your a
 jakarta-restful-ws-method-change-produces-multipart=This issue won't affect your application. \n
 jakarta-restful-ws-method-change-consumes-multipart=This issue won't affect your application. \n
 jakarta-restful-ws-method-change-context-deprecated=This issue won't affect your application. \n \
-The advice is to use CDI injection instead of @Context.
+The advise is to use CDI injection instead of @Context.
 jakarta-restful-ws-method-change-jaxblink-deprecated=This issue won't affect your application. \n \
-The advice is to stop using the inner class JaxbLink.
+The advise is to stop using the inner class JaxbLink.
 jakarta-restful-ws-method-change-jaxbadapter-deprecated=This issue won't affect your application. \n \
-The advice is to stop using the inner class JaxbAdapter.
+The advise is to stop using the inner class JaxbAdapter.
 jakarta-restful-ws-method-change-cookie-constructor-deprecated=This issue won't affect your application. \n \
-The advice is to change for Cookie.Builder class.
+The advise is to change for Cookie.Builder class.
 jakarta-restful-ws-method-change-newcookie-constructor-deprecated=This issue won't affect your application. \n \
-The advice is to change for NewCookie.Builder class.
+The advise is to change for NewCookie.Builder class.
